pyproject.toml
[project]
name = "openenv-advanced-email-triage"
version = "1.0.0"
description = "Advanced enterprise email triage OpenEnv environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.5.0",
  "transformers>=4.44.0",
  "datasets>=2.20.0",
  "sentence-transformers>=3.0.0",
  "torch>=2.4.0",
  "accelerate>=0.33.0",
  "joblib>=1.4.0",
  "sqlalchemy>=2.0.0",
  "openai>=1.40.0",
  "openenv-core>=0.2.3",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "jinja2>=3.1.0"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "server*", "graders*", "tasks*", "advanced_email_triage"]

[tool.setuptools.package-data]
advanced_email_triage = ["openenv.yaml"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
